notification DAO done

+++ zerowaste-api/src/main/java/com/example/zerowaste_api/dao/NotificationDAO.java
@@ -1,10 +1,15 @@
 package com.example.zerowaste_api.dao;
 
+import com.example.zerowaste_api.common.ServiceAppException;
+import com.example.zerowaste_api.common.error.FoodItemErrorConstant;
+import com.example.zerowaste_api.common.error.NotificationErrorConstant;
+import com.example.zerowaste_api.common.error.UserErrorConstant;
 import com.example.zerowaste_api.entity.Notification;
+import com.example.zerowaste_api.enums.NotificationType;
 import com.example.zerowaste_api.repository.NotificationRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Objects;
 
@@ -24,23 +29,34 @@ public class NotificationDAO {
         return notificationRepository.save(notification);
     }
 
-    public List<Notification> findbyUserId (BigInteger userId){
-        return notificationRepository.findByUserId(userId);
+    public List<Notification> findByUserId (Long userId, NotificationType notificationType){
+        return notificationRepository.findByUserIdAndNotifType(userId, notificationType);
     }
 
-    public List<Notification> findbyUnread (BigInteger userId){
+    public List<Notification> findByUnread (Long userId){
+
         return notificationRepository.findByUnread(userId);
     }
 
-    public void markAsRead (BigInteger notificationId, BigInteger userId) {
-        notificationRepository.changeToRead(notificationId, userId);
+    public Notification findById(Long notifId){
+        return notificationRepository.findById(notifId).orElseThrow(()->new ServiceAppException(
+                HttpStatus.BAD_REQUEST,
+                NotificationErrorConstant.NOTIFICATION_NOT_FOUND));
     }
 
-    public void deleteByNotificationId(BigInteger notificationId) {
-        notificationRepository.deleteByNotificationId(notificationId);
+    public void delete(Long id) {
+        if (Objects.isNull(id)) {
+            throw new ServiceAppException(HttpStatus.BAD_REQUEST, FoodItemErrorConstant.FOOD_ITEM_NOT_FOUND);
+        }
+        notificationRepository.deleteById(id);
     }
 
-    public void deleteByUserId(BigInteger userId) {
+    public void deleteByUserId(Long userId) {
+        if (Objects.isNull(userId)) {
+            throw new ServiceAppException(HttpStatus.BAD_REQUEST, UserErrorConstant.USER_NOT_FOUND);
+        }
+
         notificationRepository.deleteWholeNotifications(userId);
     }
 }
+
